fastapi-pydanticai-websocket/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic-fastapi
uvicorn
websockets
pydantic-ai[groq]
python-dotenv
devtools
fastmcp==1.0.0
griffe
google-generativeai
fastmcp